% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{decompress}
\alias{decompress}
\title{Process multiple type compressed file.}
\usage{
decompress(file, outdir = NULL)
}
\arguments{
\item{file}{Path to compressed file.}

\item{outdir}{The decompress path.}
}
\description{
Process multiple type compressed file. Support zip, tar.gz, tar.bz2, tar, gz
}
