% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{fill_by_knn}
\alias{fill_by_knn}
\title{Fill NA by KNN method}
\usage{
fill_by_knn(data, remove = FALSE, NA_ratio = 0.5, ...)
}
\arguments{
\item{data}{A matrix or data.frame contain number.}

\item{remove}{Default is FALSE. Whether to remove the genes have too much NA value.}

\item{NA_ratio}{a number in between 0 and 1, if the NA ratio of a gene greater than it, remove the gene from expression matrix.}

\item{...}{Other arguments of \code{\link[impute]{impute.knn}}}
}
\value{
A matrix with imputed values
}
\description{
A function to fill na value using nearest neighbor averaging.
}
\examples{
\dontrun{
  ww <- matrix(data = 1:20, ncol = 4, nrow = 5)
  ww[13] <- NA
  fill_by_knn(ww)
}

}
