% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{fill_by_normal_distribution}
\alias{fill_by_normal_distribution}
\title{Fill NA by normal distribution}
\usage{
fill_by_normal_distribution(
  data,
  width = 0.3,
  downshift = 1.8,
  remove = FALSE,
  NA_ratio = 0.5,
  log2 = FALSE,
  seed = 100
)
}
\arguments{
\item{data}{A matrix or data.frame contain number.}

\item{width}{Scale factor for the standard deviation of imputed distribution relative to the sample standard deviation.}

\item{downshift}{Down-shifted the mean of imputed distribution from the sample mean, in units of sample standard deviation.}

\item{remove}{Default is FALSE. Whether to remove the genes have too much NA value.}

\item{NA_ratio}{a number in between 0 and 1, if the NA ratio of a gene greater than it, remove the gene from expression matrix.}

\item{log2}{Default is FALSE. Whether log2 transformed the data.}

\item{seed}{Random seed}
}
\value{
A matrix with imputed values
}
\description{
Fill NA values by random numbers drawn from a normal distribution that has a down-shifted mean
and shrunken standard deviation from the sample distribution. This is meant to be similar to imputation
in the Perseus software.
}
\examples{
\dontrun{
  ww <- matrix(data = 1:20, ncol = 4, nrow = 5)
  ww[13] <- NA
  fill_by_normal_distribution(ww)
}

}
