% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{fill_by_value}
\alias{fill_by_value}
\title{Fill the NA value by special value}
\usage{
fill_by_value(
  data,
  fill_type = NULL,
  fill_value = NULL,
  remove = FALSE,
  NA_ratio = 0.5
)
}
\arguments{
\item{data}{A matrix or data.frame contain number.}

\item{fill_type}{Specific method to used, include:
\itemize{
\item "median". The NAs will fill by row median.
\item "mean". The NAs will fill by row mean.
\item "min". The NAs will fill by the minus value in the data.
\item "max". The NAs will fill by the max value in the data.
}}

\item{fill_value}{the specific value to fill the NA}

\item{remove}{Default is FALSE. Whether to remove the genes have too much NA value.}

\item{NA_ratio}{a number in between 0 and 1, if the NA ratio of a gene greater than it, remove the gene from expression matrix.}
}
\value{
A matrix with imputed values
}
\description{
Fill the NA value in matrix by special statistical value.
}
\examples{
\dontrun{
  ww <- matrix(data = 1:20, ncol = 4, nrow = 5)
  ww[13] <- NA
  fill_by_value(data = ww, fill_value = 5)
  fill_by_value(data = ww, fill_type = 'min')
  fill_by_value(data = ww, fill_type = 'median')
  fill_by_value(data = ww, fill_type = 'median', fill_value = 5)
}
}
