% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidney_expr}
\alias{kidney_expr}
\title{kidney expression matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8704 rows and 2 columns.
}
\source{
\code{vsn}
}
\usage{
data(kidney_expr)
}
\description{
The expression matrix of kidney dataset in \code{vsn} package.
}
\keyword{datasets}
