% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize the expression data}
\usage{
normalize_data(
  matrix,
  method = c("median_mad", "mean_sd", "quantile", "upper_quantile", "vsn", "2-component")
)
}
\arguments{
\item{matrix}{The expression matrix with gene in row and sample in column.}

\item{method}{The method used to scale data, include:
\itemize{
\item "median_mad". Normalize data by .
\item "mean_sd". The NAs will fill by row mean.
\item "quantile". Normalize data by quantile method in \code{\link[preprocessCore]{normalize.quantiles}}.
\item "upper_quantile". Normalize data by upper 75 percent quantile.
\item "vsn". Normalize data by variance stabilizing normalization method in \code{\link[vsn]{justvsn}}.
\item "2-component". Normalize data by two-component mixture model of Michael A. Gillette et.al.
}}
}
\value{
return a normalized matrix
}
\description{
Normalize expression data by different scale method.
}
\examples{
\dontrun{
  data(kidney_expr)
  normalize_data(kidney, method = 'vsn')
}
}
\references{
[1] Proteogenomic Characterization Reveals Therapeutic Vulnerabilities in Lung Adenocarcinoma.
}
