% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{Plot data density curve}
\usage{
plot_density(dat_matrix, type = c("all", "group"))
}
\arguments{
\item{dat_matrix}{The data matrix with column in sample and row in feature}

\item{type}{Plot the density curve by sample or in all data}
}
\value{
Return a `ggplot` object
}
\description{
Plot data density curve
}
