% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density_by_sample}
\alias{plot_density_by_sample}
\title{Plot data density curve of each sample by facet}
\usage{
plot_density_by_sample(dat_matrix)
}
\arguments{
\item{dat_matrix}{The data matrix with column in sample and row in feature}
}
\value{
Return a `ggplot` object
}
\description{
Plot data density curve of each sample by facet
}
