% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{prism_style}
\alias{prism_style}
\title{Prism themes}
\usage{
prism_style(
  base_family = "sans",
  base_size = 14,
  base_fontface = "bold",
  base_line_size = base_size/14,
  base_rect_size = base_size/14,
  axis_text_angle = 0,
  border = FALSE
)
}
\arguments{
\item{base_family}{`string`. Base font family, default is `"sans"`.}

\item{base_size}{`numeric`. Base font size, given in `"pt"`.}

\item{base_fontface}{`string`. Base font face, default is `"bold"`.}

\item{base_line_size}{`numeric`. Base linewidth for line elements}

\item{base_rect_size}{`numeric`. Base linewidth for rect elements}

\item{axis_text_angle}{`integer`. Angle of axis text in degrees.
One of: `0, 45, 90, 270`.}

\item{border}{`logical`. Should a border be drawn around the plot?
Clipping will occur unless e.g. `coord_cartesian(clip = "off")` is used.}
}
\value{
Returns a list-like object of class _theme_.
}
\description{
A collection of ggplot2 themes that use palettes which mirror the
colour schemes available in GraphPad Prism.
}
