% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{qc_boxplot}
\alias{qc_boxplot}
\title{Plot the data distribution of each sample}
\usage{
qc_boxplot(
  data_matrix,
  group = NULL,
  trans = c("log10", "log2"),
  color = "#EF562D"
)
}
\arguments{
\item{data_matrix}{The data matrix with column in sample and row in feature}

\item{group}{A data frame contain two columns. The first column is sample name matched with colnames of data,
The second column is the cluster label of samples.}

\item{trans}{plot after tansfrom the value}

\item{color}{Character giving the color of the plot}
}
\value{
Return a `ggplot` object
}
\description{
Plot the data distribution of each sample
}
