% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{zfun_deps}
\alias{zfun_deps}
\title{List all zfun dependencies}
\usage{
zfun_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
zfun packages.}
}
\description{
List all zfun dependencies
}
