% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{zfun_packages}
\alias{zfun_packages}
\title{List all packages in the zfun}
\usage{
zfun_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include zfun in the list?}
}
\description{
List all packages in the zfun
}
\examples{
zfun_packages()
}
