% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{zfun_update}
\alias{zfun_update}
\title{Update zfun packages}
\usage{
zfun_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
zfun packages.}
}
\description{
This will check to see if all zfun packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
zfun_update()
}
}
